refactor: renamed chai-blocks to blocks
