+++ jest.config.cjs
@@ -2,4 +2,5 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  testMatch: ['**/src/**/*.+(test|spec).+(ts|tsx|js)'],
 };
