+++ playwright.config.ts
@@ -10,7 +10,7 @@ import { defineConfig, devices } from "@playwright/test";
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-  testDir: "./tests",
+  testDir: "./e2e_tests",
 
   /* Run tests in files in parallel */
   fullyParallel: true,
