+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { ChaiBuilderStudio } from "./studio";
 import React from "react";
-import "./chai-blocks";
+import "./blocks";
 import "./data-providers/data.ts";
 
 const Logo: React.FC<any> = () => <h2 className="font-bold text-gray-500">Chai Builder</h2>;
